- 250403 프로젝트 일지   * 알파 블렌딩을 통한 파티클 세이더 작성   * ParticleManager 클래스 작성 시작(DrawIndexedInstanced를 바탕으로 Emitter 디버그 Cube 그리기 기능 작성)

+++ ProjectA/include/BufferMacroUtilities.h
@@ -1,3 +1,3 @@
 #pragma once
 
-#define PASS_SINGLE(a) sizeof(a), 1, &a
+#define PASS_SINGLE(a) static_cast<UINT>(sizeof(a)), 1, &a
